+++ src/main/java/mju/chatuniv/auth/application/AuthService.java
@@ -1,13 +1,12 @@
 package mju.chatuniv.auth.application;
 
-import mju.chatuniv.auth.application.dto.TokenResponse;
 import mju.chatuniv.member.application.dto.MemberCreateRequest;
 import mju.chatuniv.member.application.dto.MemberLoginRequest;
-import mju.chatuniv.member.application.dto.MemberResponse;
+import mju.chatuniv.member.domain.Member;
 
 public interface AuthService {
 
-    MemberResponse register(MemberCreateRequest memberCreateRequest);
+    Member register(MemberCreateRequest memberCreateRequest);
 
-    TokenResponse login(MemberLoginRequest memberLoginRequest);
+    String login(MemberLoginRequest memberLoginRequest);
 }
