--- src/main/java/mju/chatuniv/auth/application/dto/TokenResponse.java
@@ -1,17 +1,0 @@
-package mju.chatuniv.auth.application.dto;
-
-public class TokenResponse {
-
-    private String accessToken;
-
-    private TokenResponse() {
-    }
-
-    public TokenResponse(final String accessToken) {
-        this.accessToken = accessToken;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-}
