--- src/main/java/mju/chatuniv/auth/controller/AuthController.java
@@ -1,37 +1,0 @@
-package mju.chatuniv.auth.controller;
-
-import mju.chatuniv.auth.application.AuthService;
-import mju.chatuniv.auth.application.dto.TokenResponse;
-import mju.chatuniv.member.application.dto.MemberCreateRequest;
-import mju.chatuniv.member.application.dto.MemberLoginRequest;
-import mju.chatuniv.member.application.dto.MemberResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import javax.validation.Valid;
-
-@RequestMapping("/api/auth")
-@RestController
-public class AuthController {
-
-    private final AuthService authService;
-
-    public AuthController(final AuthService authService) {
-        this.authService = authService;
-    }
-
-    @PostMapping("/sign-up")
-    public ResponseEntity<MemberResponse> register(@RequestBody @Valid final MemberCreateRequest memberCreateRequest) {
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .body(authService.register(memberCreateRequest));
-    }
-
-    @PostMapping("/sign-in")
-    public ResponseEntity<TokenResponse> login(@RequestBody @Valid final MemberLoginRequest memberLoginRequest) {
-        return ResponseEntity.ok(authService.login(memberLoginRequest));
-    }
-}
