+++ src/main/java/mju/chatuniv/board/application/dto/BoardAllResponse.java
@@ -1,5 +1,7 @@
 package mju.chatuniv.board.application.dto;
 
+import mju.chatuniv.board.presentation.dto.BoardResponse;
+
 import java.util.List;
 
 public class BoardAllResponse {
@@ -12,7 +14,7 @@ public class BoardAllResponse {
         this.boardPageInfo = boardPageInfo;
     }
 
-    public static BoardAllResponse from(final List<BoardResponse> boards, final BoardPageInfo boardPageInfo) {
+    public static BoardAllResponse of(final List<BoardResponse> boards, final BoardPageInfo boardPageInfo) {
         return new BoardAllResponse(boards, boardPageInfo);
     }
