--- src/test/java/mju/chatuniv/comment/controller/intergration/ApplicationContextProvider.java
@@ -1,21 +1,0 @@
-package mju.chatuniv.comment.controller.intergration;
-
-import org.springframework.beans.BeansException;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ApplicationContextAware;
-import org.springframework.stereotype.Component;
-
-@Component
-public class ApplicationContextProvider implements ApplicationContextAware {
-
-    private static ApplicationContext applicationContext;
-
-    public static ApplicationContext getApplicationContext() {
-        return applicationContext;
-    }
-
-    @Override
-    public void setApplicationContext(final ApplicationContext applicationContext) throws BeansException {
-        ApplicationContextProvider.applicationContext = applicationContext;
-    }
-}
